refactor(scroll): extract scroll logic into custom hook
Move scroll detection logic from components to a reusable `useScroll` hook to improve code maintainability and reduce duplication. The hook is now used in `PageTitle` and `Supplier` components.
