+++ hooks/scroll.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+
+const useScroll = (threshold = 50) => {
+  const [isScrolled, setIsScrolled] = React.useState(false);
+
+  React.useEffect(() => {
+    const handleScroll = () => {
+      setIsScrolled(window.scrollY > threshold);
+    };
+
+    window.addEventListener("scroll", handleScroll);
+    return () => window.removeEventListener("scroll", handleScroll);
+  }, [threshold]);
+
+  return isScrolled;
+};
+
+export default useScroll;
